Some change

+++ src/main/java/com/sheroozdrive/SheroozDrive/service/FolderService.java
@@ -6,12 +6,14 @@ import com.sheroozdrive.SheroozDrive.model.Folder;
 import com.sheroozdrive.SheroozDrive.model.dto.FolderDto;
 import com.sheroozdrive.SheroozDrive.model.dto.FolderMapperTypeEnum;
 import com.sheroozdrive.SheroozDrive.model.mapper.FolderMapper;
+import com.sheroozdrive.SheroozDrive.repository.FileRepository;
 import com.sheroozdrive.SheroozDrive.repository.FolderRepository;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 @Service
@@ -19,10 +21,12 @@ public class FolderService {
 
     private final FolderRepository folderRepository;
     private final FolderMapper folderMapper;
+    private final FileRepository fileRepository;
 
-    public FolderService(FolderRepository folderRepository, FolderMapper folderMapper) {
+    public FolderService(FolderRepository folderRepository, FolderMapper folderMapper, FileRepository fileRepository) {
         this.folderRepository = folderRepository;
         this.folderMapper = folderMapper;
+        this.fileRepository = fileRepository;
     }
 
     public List<FolderDto> findByOwnerId(String ownerId) {
@@ -75,6 +79,9 @@ public class FolderService {
     }
 
     public void delete(String id) {
+        if(!folderRepository.existsById(id)){
+            throw new FolderNotFoundException(id);
+        }
         folderRepository.deleteById(id);
     }
 
@@ -82,10 +89,24 @@ public class FolderService {
         String[] items = path.split("/");
         String parentId=null;
         Folder folder=null;
-        for (String item:items) {
-            folder=folderRepository.findByNameAndParentId(item,parentId).orElseThrow(() -> new FolderNotFoundException(item));
-            parentId=folder.getId();
+
+        if (items.length>=1 && items[0].equals("root")) {
+            if(items.length==1){
+                folder=new Folder(null,"root",null,null);
+                folder.setChildFolders(folderRepository.findByParentIsNull().orElse(new ArrayList<>()));
+                folder.setFiles(fileRepository.findByFolderIsNull().orElse(new ArrayList<>()));
+            }else {
+                items = Arrays.copyOfRange(items, 1, items.length);
+                for (String item:items) {
+                    folder=folderRepository.findByNameAndParentId(item,parentId).orElseThrow(() -> new FolderNotFoundException(item));
+                    parentId=folder.getId();
+                }
+            }
+        }else{
+            throw new FolderNotFoundException(path);
         }
+
+
         return folderMapper.convertToDto(folder, FolderMapperTypeEnum.FIRST_CHILD);
     }
 }
